# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BandExplanation": {
        "additionalProperties": false,
        "description": "Answers 'why this band?' for the queue tooltip and the case header.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/PriorityBand"
          },
          "basis": {
            "description": "How the band was reached, in working language.",
            "title": "Basis",
            "type": "string"
          },
          "caps_applied": {
            "default": [],
            "description": "Guards that limited the band, e.g. text similarity alone cannot reach the top band, or an incomplete bundle lowered certainty.",
            "items": {
              "type": "string"
            },
            "title": "Caps Applied",
            "type": "array"
          }
        },
        "required": [
          "band",
          "basis"
        ],
        "title": "BandExplanation",
        "type": "object"
      },
...
      "BriefingObservation": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "kind": {
            "$ref": "#/components/schemas/ObservationKind"
          },
          "reason_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "minItems": 1,
            "title": "Source Refs",
            "type": "array"
          },
          "statement": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Statement",
            "type": "string"
          }
        },
        "required": [
          "statement",
          "kind",
          "source_refs",
          "confidence"
        ],
        "title": "BriefingObservation",
        "type": "object"
      },
      "BriefingQuestion": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "source_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "minItems": 1,
            "title": "Source Refs",
            "type": "array"
          },
          "why_it_matters": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Why It Matters",
            "type": "string"
          }
        },
        "required": [
          "question",
          "why_it_matters",
          "source_refs"
        ],
        "title": "BriefingQuestion",
        "type": "object"
      },
      "CaseBriefing": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "minimum": 1.0,
            "title": "Case Version",
            "type": "integer"
          },
          "generated_by": {
            "$ref": "#/components/schemas/GeneratedBy"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/BriefingObservation"
            },
            "maxItems": 5,
            "title": "Observations",
            "type": "array"
          },
          "open_questions": {
            "items": {
              "$ref": "#/components/schemas/BriefingQuestion"
            },
            "maxItems": 3,
            "title": "Open Questions",
            "type": "array"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "tool_calls": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ToolCallRecord"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "uncertainty_note": {
            "minLength": 1,
            "title": "Uncertainty Note",
            "type": "string"
          },
          "validation_rejected": {
            "default": false,
            "title": "Validation Rejected",
            "type": "boolean"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionStamp"
          }
        },
        "required": [
          "case_id",
          "case_version",
          "observations",
          "open_questions",
          "uncertainty_note",
          "generated_by",
          "prompt_version",
          "versions"
        ],
        "title": "CaseBriefing",
        "type": "object"
      },
      "CaseDetailResponse": {
        "additionalProperties": false,
        "description": "Everything needed to understand and disposition one case.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/BandExplanation"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "minimum": 1.0,
            "title": "Case Version",
            "type": "integer"
          },
          "comparisons": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ComparisonCandidate"
            },
            "title": "Comparisons",
            "type": "array"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "encounter_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter End"
          },
          "encounter_start": {
            "format": "date-time",
            "title": "Encounter Start",
            "type": "string"
          },
          "evidence_completeness": {
            "$ref": "#/components/schemas/EvidenceCompleteness"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ClaimLineView"
            },
            "title": "Lines",
            "type": "array"
          },
          "participant_token": {
            "title": "Participant Token",
            "type": "string"
          },
          "primary_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_token": {
            "title": "Provider Token",
            "type": "string"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonDto"
            },
            "title": "Reasons",
            "type": "array"
          },
          "sources": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceResource"
            },
            "title": "Sources",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the system would suggest, e.g. REQUEST_EVIDENCE when the bundle is thin. A suggestion only \u2014 the reviewer chooses, and must still supply a reason.",
            "title": "Suggested Action"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "title": "Timeline",
            "type": "array"
          },
          "total_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionStamp"
          }
        },
        "required": [
          "case_id",
          "case_version",
          "state",
          "participant_token",
          "provider_token",
          "total_amount",
          "currency",
          "encounter_start",
          "encounter_end",
          "primary_reason",
          "reasons",
          "band",
          "lines",
          "timeline",
          "evidence_completeness",
          "versions"
        ],
        "title": "CaseDetailResponse",
        "type": "object"
      },
...
      "CaseState": {
        "description": "Functional state model from the master plan \u00a7 10.",
        "enum": [
          "NEW",
          "SCREENED",
          "IN_REVIEW",
          "EVIDENCE_REQUESTED",
          "DISMISSED",
          "CONFIRMED_ANOMALY",
          "ESCALATED",
          "INVALID_INPUT"
        ],
        "title": "CaseState",
        "type": "string"
      },
...
      "ClaimLineView": {
        "additionalProperties": false,
        "description": "One billed line with its support state.\n\n`NOT_ASSESSABLE` is distinct from `UNSUPPORTED`: the first means the record was too thin\nto judge, the second means the evidence is genuinely absent. Conflating them is the\ndifference between requesting a document and alleging a service never happened.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "line_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Amount",
            "type": "string"
          },
          "line_id": {
            "title": "Line Id",
            "type": "string"
          },
          "quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "type": "string"
          },
          "service_at": {
            "format": "date-time",
            "title": "Service At",
            "type": "string"
          },
          "support_state": {
            "description": "SUPPORTED | PARTIALLY_SUPPORTED | UNSUPPORTED | NOT_ASSESSABLE",
            "title": "Support State",
            "type": "string"
          }
        },
        "required": [
          "line_id",
          "code",
          "description",
          "quantity",
          "line_amount",
          "service_at",
          "support_state"
        ],
        "title": "ClaimLineView",
        "type": "object"
      },
      "ComparisonCandidate": {
        "additionalProperties": false,
        "description": "Side-by-side pair for repeat-billing and cloned-documentation reasons.",
        "properties": {
          "candidate_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Case Id"
          },
          "candidate_claim_id": {
            "title": "Candidate Claim Id",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ComparisonField"
            },
            "title": "Fields",
            "type": "array"
          },
          "overlap_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap End"
          },
          "overlap_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap Start"
          },
          "similarity_components": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Similarity Components",
            "type": "object"
          },
          "template_caveat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present on clone reasons. Legitimate template-based documentation produces high similarity too, and the reviewer must read that before acting.",
            "title": "Template Caveat"
          }
        },
        "required": [
          "candidate_claim_id",
          "fields"
        ],
        "title": "ComparisonCandidate",
        "type": "object"
      },
      "ComparisonField": {
        "additionalProperties": false,
        "properties": {
          "field_name": {
            "title": "Field Name",
            "type": "string"
          },
          "left_value": {
            "title": "Left Value",
            "type": "string"
          },
          "matches": {
            "title": "Matches",
            "type": "boolean"
          },
          "right_value": {
            "title": "Right Value",
            "type": "string"
          }
        },
        "required": [
          "field_name",
          "left_value",
          "right_value",
          "matches"
        ],
        "title": "ComparisonField",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "Stable identifiers. Append new members; never repurpose an existing one.",
        "enum": [
          "BUNDLE_TOO_LARGE",
          "BUNDLE_UNSUPPORTED_CONTENT_TYPE",
          "BUNDLE_DEPTH_EXCEEDED",
          "BUNDLE_MALFORMED_JSON",
          "BUNDLE_SCHEMA_INVALID",
          "BUNDLE_UNKNOWN_RESOURCE_TYPE",
          "BUNDLE_DANGLING_REFERENCE",
          "BUNDLE_CIRCULAR_REFERENCE",
          "BUNDLE_DUPLICATE_RESOURCE_ID",
          "BUNDLE_TOTAL_MISMATCH",
          "INGESTION_NOT_FOUND",
          "CASE_NOT_FOUND",
          "EVALUATION_RUN_NOT_FOUND",
          "BUNDLE_NOT_SCREENABLE",
          "DISPOSITION_REASON_REQUIRED",
          "CASE_VERSION_CONFLICT",
          "DISPOSITION_INVALID_TRANSITION",
          "AUDIT_FORBIDDEN",
          "CASE_ACCESS_FORBIDDEN",
          "CASE_REOPEN_FORBIDDEN",
          "USER_MANAGEMENT_FORBIDDEN",
          "SESSION_INVALID_CREDENTIALS",
          "SESSION_ACCOUNT_DEACTIVATED",
          "USER_NOT_FOUND",
          "USER_SELF_MODIFICATION_REFUSED",
          "USER_NO_CHANGE_REQUESTED",
          "BRIEFING_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The single error envelope for every endpoint.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceCompleteness": {
        "additionalProperties": false,
        "description": "How much of the expected supporting evidence was actually present.\n\nSurfaced in the queue so a reviewer can tell a thin record from a suspicious one before\nopening the case.",
        "properties": {
          "bundle_complete": {
            "title": "Bundle Complete",
            "type": "boolean"
          },
          "missing_reference_count": {
            "minimum": 0.0,
            "title": "Missing Reference Count",
            "type": "integer"
          },
          "supported_lines": {
            "minimum": 0.0,
            "title": "Supported Lines",
            "type": "integer"
          },
          "total_lines": {
            "minimum": 0.0,
            "title": "Total Lines",
            "type": "integer"
          }
        },
        "required": [
          "supported_lines",
          "total_lines",
          "missing_reference_count",
          "bundle_complete"
        ],
        "title": "EvidenceCompleteness",
        "type": "object"
      },
      "EvidenceRefDto": {
        "additionalProperties": false,
        "description": "A pointer the UI must be able to open.\n\nA reference that does not resolve is a defect, not an empty panel \u2014 see\n`docs/canonical/06_evaluation_plan.md` \u00a7 Evidence-reference validity.",
        "properties": {
          "label": {
            "description": "Short working-language description of the resource.",
            "title": "Label",
            "type": "string"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          }
        },
        "required": [
          "resource_type",
          "resource_id",
          "label"
        ],
        "title": "EvidenceRefDto",
        "type": "object"
      },
      "GeneratedBy": {
        "enum": [
          "LLM",
          "TEMPLATE"
        ],
        "title": "GeneratedBy",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PriorityBand": {
        "description": "Queue ordering bands.\n\n`NO_OBSERVED_RISK` means no selected detector fired at this engine version. It must never\nbe rendered as \"clean\" or \"safe\" \u2014 the system is not entitled to that claim.",
        "enum": [
          "DETERMINISTIC_CONFLICT",
          "HIGH_PRIORITY_SIGNAL",
          "NEEDS_CONTEXT",
          "NO_OBSERVED_RISK"
        ],
        "title": "PriorityBand",
        "type": "string"
      },
...
      "ReasonCode": {
        "description": "Stable identifiers used by the UI and by tests. Never renumbered.",
        "enum": [
          "LINE_WITHOUT_COMPLETED_PROCEDURE",
          "LINE_WITHOUT_MEDICATION_DISPENSE",
          "SUPPORTING_EVIDENCE_ENTERED_IN_ERROR",
          "OVERLAPPING_CLAIM_SAME_EPISODE",
          "DUPLICATE_CLAIM_FINGERPRINT",
          "NEAR_DUPLICATE_DOCUMENTATION",
          "EPISODE_SPLIT_ACROSS_CLAIMS"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
      "ReasonDto": {
        "additionalProperties": false,
        "description": "One reason, with its evidence and whatever argues against it.\n\n`counter_evidence` is a first-class field rather than a separate lookup: the reviewer must\nsee what weakens a signal on the same screen as the signal itself.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ReasonCode"
          },
          "component_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Per-component scores, kept alongside the reason rather than collapsed.",
            "title": "Component Scores",
            "type": "object"
          },
          "counter_evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "title": "Counter Evidence",
            "type": "array"
          },
          "counter_evidence_notes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CounterEvidenceDto"
            },
            "title": "Counter Evidence Notes",
            "type": "array"
          },
          "deterministic": {
            "title": "Deterministic",
            "type": "boolean"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "title": "Evidence",
            "type": "array"
          },
          "expected_support": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceType"
            },
            "title": "Expected Support",
            "type": "array"
          },
          "mode": {
            "$ref": "#/components/schemas/RiskMode"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "sentence": {
            "description": "From the reason catalog. Clients never compose their own.",
            "title": "Sentence",
            "type": "string"
          }
        },
        "required": [
          "code",
          "mode",
          "sentence",
          "deterministic",
          "evidence",
          "ruleset_version"
        ],
        "title": "ReasonDto",
        "type": "object"
      },
...
      "ResourceType": {
        "enum": [
          "Claim",
          "ClaimLine",
          "Encounter",
          "Condition",
          "Procedure",
          "Medication",
          "Diagnostic",
          "Document",
          "Account",
          "ChargeItem",
          "Invoice",
          "Episode",
          "Practitioner"
        ],
        "title": "ResourceType",
        "type": "string"
      },
...
      "SourceResource": {
        "additionalProperties": false,
        "description": "The resource behind an evidence reference, so every reference opens.\n\nAn unresolvable reference is recorded here with `MISSING` rather than dropped. Dropping it\nwould render as a shorter list \u2014 indistinguishable from a reason that cited less \u2014 which is\nexactly how a broken evidence trail hides.",
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/SourceAvailability"
          },
          "fields": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceField"
            },
            "title": "Fields",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          }
        },
        "required": [
          "resource_type",
          "resource_id",
          "label",
          "availability"
        ],
        "title": "SourceResource",
        "type": "object"
      },
      "TimelineEvent": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "resource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceRefDto"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "occurred_at",
          "kind",
          "label"
        ],
        "title": "TimelineEvent",
        "type": "object"
      },
      "ToolCallRecord": {
        "additionalProperties": false,
        "description": "One function the briefing chose to read. The log *is* the transparency artifact.",
        "properties": {
          "arguments": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Arguments",
            "type": "object"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "tool"
        ],
        "title": "ToolCallRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "additionalProperties": false,
        "description": "One problem found in a submitted bundle, pointed at a specific resource.\n\n`resource_type` and `resource_id` are what make the ingest screen's error list\nactionable \u2014 the operator learns *which* resource to fix, not merely that something is\nwrong somewhere.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "description": "Human-readable explanation. Never contains medical text.",
            "title": "Detail",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
      "VersionStamp": {
        "additionalProperties": false,
        "description": "Engine identity carried on every screening result and audit event.",
        "properties": {
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "ruleset_version",
          "engine_version",
          "dataset_version"
        ],
        "title": "VersionStamp",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Claim evidence integrity screening. Synthetic data only.",
    "title": "TilikKlaim API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases/{case_id}": {
      "get": {
        "description": "Claim lines, reasons with evidence and counter-evidence, timeline, and comparisons.",
        "operationId": "get_case_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "reviewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Case detail",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/v1/cases/{case_id}/briefing": {
      "get": {
        "description": "Observations with source references, open questions, and an uncertainty note.\n\nNever a score, a band, or a state transition. Off by default; the deterministic template is\nthe answer when no model is configured.",
        "operationId": "get_briefing_v1_cases__case_id__briefing_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": true,
              "title": "Stream",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "reviewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseBriefing"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read-only case briefing (non-authoritative)",
        "tags": [
          "briefing"
        ]
      }
    },
...
  }
}